+++ ICE/Ui/DebugWindowTabs/PlayerInfo.cs
@@ -50,7 +50,7 @@ namespace ICE.Ui.DebugWindowTabs
             ImGui.Text($"Items on person: ");
             foreach (var item in ConsumableInfo.Food)
             {
-                if (PlayerHelper.GetItemCount((int)item.Id, out var count) && count > 0)
+                if (PlayerHelper.GetItemCount(item.Id, out var count) && count > 0)
                     ImGui.Text($"{item.Name} | {item.Id}");
             }
         }
